cycleAI/requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-huggingface
duckduckgo-search
chromadb
beautifulsoup4
requests
langgraph
streamlit
wikipedia
python-dotenv
pandas
transformers
torch
sentence-transformers
playwright
lxml
